root_requirements.txt
appwrite
nltk
numpy
python-dotenv
flask
flask-cors
